Added ticket scanning, admin's view of EventApproveals
Every User on the Profile can tap on MyID button which displays users ID in QR code. That QR code can be used by Exhibitor to check if user bought tickets for event. Exhibitor gets clear and simple green or red screen for simple checking. For QR code scanning, MercariQRScanner pod was used. On EventApproveal, ApprovealStatus enum was added for easier checks.  On AdminDashboard, Oranize button was added which displays table of all EventApproveals. Only those approveals that are in awating feedback status can be changed. For others, details view is available.